Create Quiz App
This is a productive app for building quizzes for people be it for trivia, technical, medical, or academic.  I credit Angela Yu from London App Brewery for her tutorial.  The code embodies a model file with question banks, view controller, and an imported library from older versions of objective c which has been included in the code.  The top is the View Controller and below is the Question Bank file in Swift under the Model folder.